refactor(utils): reorganize createComponentProps utility
