+++ packages/react/src/utils/create-component-props/index.ts
@@ -1,0 +1,3 @@
+import createComponentProps from "./create-component-props";
+
+export default createComponentProps;
