+++ packages/react/src/utils/index.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import-x/prefer-default-export */
+export { default as createComponentProps } from "./create-component-props";
